[life improvement] Moving Python files around  (#4531)
* Moved components to the tf folder and moved the TrainerFactory to the `trainer` folder

* Addressing comments

* Editing the migrating doc

* fixing test
