+++ ml-agents/mlagents/trainers/tests/test_trainer_util.py
@@ -3,13 +3,14 @@ import io
 import os
 from unittest.mock import patch
 
-from mlagents.trainers import trainer_util
+from mlagents.trainers.trainer import TrainerFactory
 from mlagents.trainers.cli_utils import load_config, _load_config
 from mlagents.trainers.ppo.trainer import PPOTrainer
 from mlagents.trainers.exception import TrainerConfigError, UnityTrainerException
 from mlagents.trainers.settings import RunOptions
 from mlagents.trainers.tests.dummy_config import ppo_dummy_config
 from mlagents.trainers.environment_parameter_manager import EnvironmentParameterManager
+from mlagents.trainers.directory_utils import validate_existing_directories
 
 
 @pytest.fixture
@@ -49,7 +50,7 @@ def test_initialize_ppo_trainer(BehaviorSpecMock, dummy_config):
         assert artifact_path == os.path.join(output_path, brain_name)
 
     with patch.object(PPOTrainer, "__init__", mock_constructor):
-        trainer_factory = trainer_util.TrainerFactory(
+        trainer_factory = TrainerFactory(
             trainer_config=base_config,
             output_path=output_path,
             train_model=train_model,
@@ -71,7 +72,7 @@ def test_handles_no_config_provided():
     brain_name = "testbrain"
     no_default_config = RunOptions().behaviors
 
-    trainer_factory = trainer_util.TrainerFactory(
+    trainer_factory = TrainerFactory(
         trainer_config=no_default_config,
         output_path="output_path",
         train_model=True,
@@ -112,25 +113,25 @@ you:
 def test_existing_directories(tmp_path):
     output_path = os.path.join(tmp_path, "runid")
     # Test fresh new unused path - should do nothing.
-    trainer_util.handle_existing_directories(output_path, False, False)
+    validate_existing_directories(output_path, False, False)
     # Test resume with fresh path - should throw an exception.
     with pytest.raises(UnityTrainerException):
-        trainer_util.handle_existing_directories(output_path, True, False)
+        validate_existing_directories(output_path, True, False)
 
     # make a directory
     os.mkdir(output_path)
     # Test try to train w.o. force, should complain
     with pytest.raises(UnityTrainerException):
-        trainer_util.handle_existing_directories(output_path, False, False)
+        validate_existing_directories(output_path, False, False)
     # Test try to train w/ resume - should work
-    trainer_util.handle_existing_directories(output_path, True, False)
+    validate_existing_directories(output_path, True, False)
     # Test try to train w/ force - should work
-    trainer_util.handle_existing_directories(output_path, False, True)
+    validate_existing_directories(output_path, False, True)
 
     # Test initialize option
     init_path = os.path.join(tmp_path, "runid2")
     with pytest.raises(UnityTrainerException):
-        trainer_util.handle_existing_directories(output_path, False, True, init_path)
+        validate_existing_directories(output_path, False, True, init_path)
     os.mkdir(init_path)
     # Should pass since the directory exists now.
-    trainer_util.handle_existing_directories(output_path, False, True, init_path)
+    validate_existing_directories(output_path, False, True, init_path)
