+++ ml-agents/mlagents/trainers/tf/components/reward_signals/curiosity/model.py
