+++ ml-agents/mlagents/trainers/tf/components/reward_signals/curiosity/signal.py
@@ -2,8 +2,13 @@ from typing import Any, Dict
 import numpy as np
 from mlagents.tf_utils import tf
 
-from mlagents.trainers.components.reward_signals import RewardSignal, RewardSignalResult
-from mlagents.trainers.components.reward_signals.curiosity.model import CuriosityModel
+from mlagents.trainers.tf.components.reward_signals import (
+    RewardSignal,
+    RewardSignalResult,
+)
+from mlagents.trainers.tf.components.reward_signals.curiosity.model import (
+    CuriosityModel,
+)
 from mlagents.trainers.policy.tf_policy import TFPolicy
 from mlagents.trainers.buffer import AgentBuffer
 from mlagents.trainers.settings import CuriositySettings
