+++ ml-agents/mlagents/trainers/tf/components/reward_signals/gail/model.py
