+++ ml-agents/mlagents/trainers/trainer/__init__.py
@@ -1,1 +1,2 @@
 from mlagents.trainers.trainer.trainer import Trainer  # noqa
+from mlagents.trainers.trainer.trainer_factory import TrainerFactory  # noqa
